+++ sdk/storagesync/arm-storagesync/src/models/mappers.ts
@@ -175,6 +175,17 @@ export const OperationResourceMetricSpecification: coreClient.CompositeMapper = 
           name: "String"
         }
       },
+      supportedAggregationTypes: {
+        serializedName: "supportedAggregationTypes",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      },
       fillGapWithZero: {
         serializedName: "fillGapWithZero",
         type: {
@@ -277,8 +288,8 @@ export const StorageSyncApiError: coreClient.CompositeMapper = {
           className: "StorageSyncErrorDetails"
         }
       },
-      innerError: {
-        serializedName: "innerError",
+      innererror: {
+        serializedName: "innererror",
         type: {
           name: "Composite",
           className: "StorageSyncInnerErrorDetails"
@@ -529,6 +540,58 @@ export const Resource: coreClient.CompositeMapper = {
         readOnly: true,
         type: {
           name: "String"
+        }
+      },
+      systemData: {
+        serializedName: "systemData",
+        type: {
+          name: "Composite",
+          className: "SystemData"
+        }
+      }
+    }
+  }
+};
+
+export const SystemData: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "SystemData",
+    modelProperties: {
+      createdBy: {
+        serializedName: "createdBy",
+        type: {
+          name: "String"
+        }
+      },
+      createdByType: {
+        serializedName: "createdByType",
+        type: {
+          name: "String"
+        }
+      },
+      createdAt: {
+        serializedName: "createdAt",
+        type: {
+          name: "DateTime"
+        }
+      },
+      lastModifiedBy: {
+        serializedName: "lastModifiedBy",
+        type: {
+          name: "String"
+        }
+      },
+      lastModifiedByType: {
+        serializedName: "lastModifiedByType",
+        type: {
+          name: "String"
+        }
+      },
+      lastModifiedAt: {
+        serializedName: "lastModifiedAt",
+        type: {
+          name: "DateTime"
         }
       }
     }
@@ -1094,6 +1157,29 @@ export const TriggerChangeDetectionParameters: coreClient.CompositeMapper = {
   }
 };
 
+export const CloudEndpointAfsShareMetadataCertificatePublicKeys: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "CloudEndpointAfsShareMetadataCertificatePublicKeys",
+    modelProperties: {
+      firstKey: {
+        serializedName: "firstKey",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      },
+      secondKey: {
+        serializedName: "secondKey",
+        readOnly: true,
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const ServerEndpointSyncStatus: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -1507,6 +1593,13 @@ export const ServerEndpointCloudTieringStatus: coreClient.CompositeMapper = {
           name: "Composite",
           className: "CloudTieringDatePolicyStatus"
         }
+      },
+      lowDiskMode: {
+        serializedName: "lowDiskMode",
+        type: {
+          name: "Composite",
+          className: "CloudTieringLowDiskMode"
+        }
       }
     }
   }
@@ -1746,6 +1839,29 @@ export const CloudTieringDatePolicyStatus: coreClient.CompositeMapper = {
         readOnly: true,
         type: {
           name: "DateTime"
+        }
+      }
+    }
+  }
+};
+
+export const CloudTieringLowDiskMode: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "CloudTieringLowDiskMode",
+    modelProperties: {
+      lastUpdatedTimestamp: {
+        serializedName: "lastUpdatedTimestamp",
+        readOnly: true,
+        type: {
+          name: "DateTime"
+        }
+      },
+      state: {
+        serializedName: "state",
+        readOnly: true,
+        type: {
+          name: "String"
         }
       }
     }
@@ -2505,6 +2621,7 @@ export const ServerEndpointCreateParameters: coreClient.CompositeMapper = {
         }
       },
       tierFilesOlderThanDays: {
+        defaultValue: 0,
         constraints: {
           InclusiveMaximum: 2147483647,
           InclusiveMinimum: 0
@@ -3577,6 +3694,27 @@ export const CloudEndpointsTriggerChangeDetectionHeaders: coreClient.CompositeMa
           name: "String"
         }
       },
+      xMsRequestId: {
+        serializedName: "x-ms-request-id",
+        type: {
+          name: "String"
+        }
+      },
+      xMsCorrelationRequestId: {
+        serializedName: "x-ms-correlation-request-id",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const CloudEndpointsAfsShareMetadataCertificatePublicKeysHeaders: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "CloudEndpointsAfsShareMetadataCertificatePublicKeysHeaders",
+    modelProperties: {
       xMsRequestId: {
         serializedName: "x-ms-request-id",
         type: {